Create /artist/X page route. Sorta hacky
I'm committing this to a branch because it's not ready for main- there
are some inconsistencies in routing, like when you're looking at an
address, click back, or when you have a card open, click the X, etc...
These scenarios are buggy.

I probably should instead refactor the code to handle
artist navigation better to get this change done.